add transaction fee

+++ src/AElfScanServer.Token/Dtos/TokenCommonDto.cs
@@ -3,8 +3,8 @@ namespace AElfScanServer.Token.Dtos;
 public class TokenCommonDto
 {
     public TokenBaseInfo Token { get; set; }
-    public long TotalSupply { get; set; }
-    public long CirculatingSupply { get; set; }
+    public decimal TotalSupply { get; set; }
+    public decimal CirculatingSupply { get; set; }
     public long Holders { get; set; }
     public double HolderPercentChange24H { get; set; }
     public long TransferCount { get; set; }
